+++ Changelog.md
@@ -260,3 +260,7 @@ Since v2.5:
 - Removed `expect` values: `auto`, `primitive`.
 - `expect` now defaults to accepting every data type.
 - Added a `multiline_handling` option for regex. Changed the default to return the first matching line only.
+- All VCIT registries now default to the `variants-cit` namespace.
+
+## TODO
+- `maches_any` and `matches_all` simplified keys.